# perf: use `get_cached_value` instead of `db.get_value` in setup module