Continues: #32836

Replaced `db.get_value` with `get_cached_value` in **setup** Module.

### Before
```
In [1]: %timeit doc.validate_default_accounts()
7.33 ms ± 61.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

### After
```
In [1]: %timeit doc.validate_default_accounts()
40.2 µs ± 3.78 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```